Added icon for SubAppTypes
Also moved LabelProvider for SubAppHierarchyDialog
