+++ plugins/org.eclipse.fordiac.ide.application/src/org/eclipse/fordiac/ide/application/utilities/SubAppHierarchyDialog.java
@@ -19,17 +19,24 @@ import java.util.List;
 import org.eclipse.emf.ecore.EObject;
 import org.eclipse.emf.ecore.util.EcoreUtil;
 import org.eclipse.fordiac.ide.application.Messages;
-import org.eclipse.fordiac.ide.application.handlers.MoveThroughHierarchyHandler.TreeNodeLabelProvider;
 import org.eclipse.fordiac.ide.model.libraryElement.Application;
 import org.eclipse.fordiac.ide.model.libraryElement.AutomationSystem;
 import org.eclipse.fordiac.ide.model.libraryElement.CompositeFBType;
 import org.eclipse.fordiac.ide.model.libraryElement.FBNetwork;
 import org.eclipse.fordiac.ide.model.libraryElement.FBNetworkElement;
+import org.eclipse.fordiac.ide.model.libraryElement.INamedElement;
 import org.eclipse.fordiac.ide.model.libraryElement.SubApp;
+import org.eclipse.fordiac.ide.model.libraryElement.SubAppType;
+import org.eclipse.fordiac.ide.model.libraryElement.TypedSubApp;
 import org.eclipse.fordiac.ide.model.libraryElement.UntypedSubApp;
+import org.eclipse.fordiac.ide.ui.imageprovider.FordiacImage;
+import org.eclipse.jface.viewers.DelegatingStyledCellLabelProvider.IStyledLabelProvider;
+import org.eclipse.jface.viewers.LabelProvider;
+import org.eclipse.jface.viewers.StyledString;
 import org.eclipse.jface.viewers.TreeNode;
 import org.eclipse.jface.viewers.TreeNodeContentProvider;
 import org.eclipse.jface.viewers.TreeViewer;
+import org.eclipse.swt.graphics.Image;
 import org.eclipse.swt.widgets.Composite;
 import org.eclipse.swt.widgets.Display;
 import org.eclipse.ui.dialogs.ElementTreeSelectionDialog;
@@ -96,5 +103,40 @@ public class SubAppHierarchyDialog {
 			}
 		});
 		parent.setChildren(nodeList.toArray(new TreeNode[0]));
+	}
+
+	private static class TreeNodeLabelProvider extends LabelProvider implements IStyledLabelProvider {
+
+		@Override
+		public String getText(final Object element) {
+			if (element instanceof final TreeNode treeNode && treeNode.getValue() instanceof final INamedElement e) {
+				return e.getName();
+			}
+			return element.toString();
+		}
+
+		@Override
+		public Image getImage(final Object element) {
+			if (element instanceof final TreeNode node) {
+				if (node.getValue() instanceof UntypedSubApp) {
+					return FordiacImage.ICON_SUB_APP.getImage();
+				}
+				if (node.getValue() instanceof TypedSubApp) {
+					return FordiacImage.ICON_SUB_APP.getImage();
+				}
+				if (node.getValue() instanceof SubAppType) {
+					return FordiacImage.ICON_SUB_APP_TYPE.getImage();
+				}
+				if (node.getValue() instanceof Application) {
+					return FordiacImage.ICON_APPLICATION.getImage();
+				}
+			}
+			return super.getImage(element);
+		}
+
+		@Override
+		public StyledString getStyledText(final Object element) {
+			return null;
+		}
 	}
 }
